KOGITO-216 Rename into kogito-dmn
