--- drools/kie-dmn/kogito-dmn-rest-utils/.gitignore
@@ -1,11 +1,0 @@
-/target
-/local
-/bin
-
-# Eclipse, Netbeans and IntelliJ files
-/.*
-!.gitignore
-/nbproject
-/*.ipr
-/*.iws
-/*.iml
