--- drools/kie-dmn/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.kie.kogito</groupId>
-    <artifactId>drools</artifactId>
-    <version>8.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>kogito-dmn</artifactId>
-  <packaging>pom</packaging>
-
-  <name>Kogito :: Decision Model Notation</name>
-  <description>Kogito Top level pom for the DMN support implementation</description>
-
-  <modules>
-    <module>kogito-dmn-rest-utils</module>
-  </modules>
-
-</project>
