+++ drools/kogito-dmn/pom.xml
@@ -1,27 +1,23 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-dmn</artifactId>
+    <artifactId>drools</artifactId>
     <version>8.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-dmn-rest-utils</artifactId>
-  <description>Kogito DMN REST utilities</description>
 
-  <properties>
-    <surefire-plugin.version>2.22.0</surefire-plugin.version>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-  </properties>
+  <artifactId>kogito-dmn</artifactId>
+
+  <name>Kogito :: Decision Model Notation</name>
+  <description>Kogito wrapper for Kie Drools DMN open source engine</description>
 
   <dependencies>
     <dependency>
       <groupId>javax.json.bind</groupId>
       <artifactId>javax.json.bind-api</artifactId>
-      <version>1.0</version>
     </dependency>
     <dependency>
       <groupId>org.kie</groupId>
@@ -118,16 +114,9 @@
       <plugin>
         <groupId>org.jboss.jandex</groupId>
         <artifactId>jandex-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>make-index</id>
-            <goals>
-              <goal>jandex</goal>
-            </goals>
-          </execution>
-        </executions>
       </plugin>
     </plugins>
   </build>
 
 </project>
+
