+++ drools/kogito-dmn/src/main/java/org/kie/kogito/dmn/rest/DMNInputDataInfo.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.dmn.kogito.rest.quarkus;
+package org.kie.kogito.dmn.rest;
 
 import java.io.Serializable;
