+++ drools/kogito-dmn/src/main/java/org/kie/kogito/dmn/rest/DMNUnaryTestsInfo.java
@@ -14,45 +14,36 @@
  * limitations under the License.
  */
 
-package org.kie.dmn.kogito.rest.quarkus;
+package org.kie.kogito.dmn.rest;
 
 import java.io.Serializable;
 
 import javax.json.bind.annotation.JsonbProperty;
 
-import org.kie.dmn.api.core.ast.DecisionNode;
+public class DMNUnaryTestsInfo implements Serializable {
 
-public class DMNDecisionInfo implements Serializable {
+    private String text;
+    private String expressionLanguage;
 
-    private String id;
-    private String name;
-
-    public DMNDecisionInfo() {
+    public DMNUnaryTestsInfo() {
         // Intentionally blank.
     }
 
-    public static DMNDecisionInfo of(DecisionNode decisionNode) {
-        DMNDecisionInfo res = new DMNDecisionInfo();
-        res.setName(decisionNode.getName());
-        res.setId(decisionNode.getId());
-        return res;
+    @JsonbProperty("unarytests-text")
+    public String getText() {
+        return text;
     }
 
-    @JsonbProperty("decision-id")
-    public String getId() {
-        return id;
+    public void setText(String text) {
+        this.text = text;
     }
 
-    public void setId(String id) {
-        this.id = id;
+    @JsonbProperty("unarytests-expressionLanguage")
+    public String getExpressionLanguage() {
+        return expressionLanguage;
     }
 
-    @JsonbProperty("decision-name")
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
+    public void setExpressionLanguage(String expressionLanguage) {
+        this.expressionLanguage = expressionLanguage;
     }
 }
