+++ drools/pom.xml
@@ -29,7 +29,7 @@
     <module>drools-core</module>
     <module>drools-compiler</module>
     <module>drools-model</module>
-    <module>kie-dmn</module>
+    <module>kogito-dmn</module>
     <module>drools-decisiontables</module>
   </modules>
