+++ kogito-codegen/pom.xml
@@ -69,7 +69,7 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-dmn-rest-utils</artifactId> <!-- depend on the utility wrapper module to avoid repeating dependency exclusions -->
+      <artifactId>kogito-dmn</artifactId> <!-- depend on the utility wrapper module to avoid repeating dependency exclusions -->
     </dependency>
 
     <!-- image metadata handling -->
