+++ kogito-codegen/src/main/resources/class-templates/DMNApplicationClassDeclTemplate.java
@@ -1,7 +1,7 @@
 
 public class DecisionModels implements org.kie.kogito.decision.DecisionModels {
 
-    static org.kie.dmn.api.core.DMNRuntime dmnRuntime = org.kie.dmn.kogito.rest.quarkus.DMNKogitoQuarkus.createGenericDMNRuntime();
+    static org.kie.dmn.api.core.DMNRuntime dmnRuntime = org.kie.kogito.dmn.DMNKogito.createGenericDMNRuntime();
 
     public org.kie.kogito.decision.DecisionModel getDecisionModel(java.lang.String namespace, java.lang.String name) {
         return new org.kie.kogito.decision.DecisionModel() {
